pyproject.toml
[project]
name = "editardeltalake"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "deltalake>=1.1.2",
    "duckdb>=1.3.2",
    "ibis-framework>=10.6.0",
    "marimo>=0.14.13",
]
